finalrequirements.txt
streamlit>=1.28.0
easyocr>=1.7.0
textblob>=0.17.1
Pillow>=10.0.0
torch>=2.0.0
numpy>=1.24.0
opencv-python-headless>=4.8.0
matplotlib>=3.7.0
requests>=2.31.0
